requirements.txt
django==2.2.6
djangorestframework
django-cors-headers
mysqlclient
django-environ
